Move Rectangle to rectangle.h
Actually, there is nothing X-specific in Rectangle, and moreover, it is
used in entirely X-independent code, so move it to its own module, which
already existed.

I didn't adjust any of the code moved from rectangle.cpp. I've only
removed two helper functions rects_intersect() and intersection_area()
which are already part of the Rectangle class.
