+++ src/floating.h
@@ -2,7 +2,7 @@
 #define __HERBST_FLOATING_H_
 
 #include "types.h"
-#include "x11-types.h"
+#include "rectangle.h"
 
 class Client;
 class HSTag;
