+++ src/monitordetection.h
@@ -3,7 +3,7 @@
 #include <string>
 #include <vector>
 
-#include "x11-types.h"
+#include "rectangle.h"
 
 class XConnection;
