+++ src/panelmanager.h
@@ -3,8 +3,8 @@
 #include <X11/X.h>
 #include <unordered_map>
 
+#include "rectangle.h"
 #include "signal.h"
-#include "x11-types.h"
 
 class Panel;
 class Settings;
